+++ tests/run_test.py
@@ -1,5 +1,5 @@
 import os
-os.environ.setdefault('BUEM_WEATHER_DIR', os.path.abspath('src/buem/data'))
+os.environ.setdefault('BUEM_WEATHER_DIR', os.path.abspath('src/buem/data/weather'))
 
 from buem.config.cfg_attribute import cfg
 from buem.thermal.model_buem import ModelBUEM